requirements.txt
impacket
ftplib
libnfs
python-magic